changes to check first if record exists

+++ appsrc/modules/files/controllers/fileCategoryController.js
@@ -50,18 +50,35 @@ exports.getFileCategories = async (req, res, next) => {
 };
 
 exports.deleteFileCategory = async (req, res, next) => {
-  var obj = {};
-  obj.query = {
-    _id: req.params.id,
-    isArchived: true
-  }
+
+  const { FileCategory } = require('../models');
   try {
-    const result = await this.dbservice.deleteObjectAfterVerification(FileCategory, obj);
-    res.status(StatusCodes.OK).send(rtnMsg.recordDelMessage(StatusCodes.OK, result));
+    const response = await this.dbservice.getObjectList(FileCategory, this.fields, {category: req.params.id}, this.orderBy, this.populate);
+
+
+    if(response.length == 0) {
+      var obj = {};
+      obj.query = {
+        _id: req.params.id,
+        isArchived: true
+      }
+      try {
+        const result = await this.dbservice.deleteObjectAfterVerification(FileCategory, obj);
+        res.status(StatusCodes.OK).send(rtnMsg.recordDelMessage(StatusCodes.OK, result));
+      } catch (error) {
+        logger.error(new Error(error));
+        res.status(StatusCodes.INTERNAL_SERVER_ERROR).send(getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR));
+      }
+    } else {
+      res.status(StatusCodes.CONFLICT).send(rtnMsg.recordDelMessage(StatusCodes.CONFLICT, null));
+    }
   } catch (error) {
     logger.error(new Error(error));
     res.status(StatusCodes.INTERNAL_SERVER_ERROR).send(getReasonPhrase(StatusCodes.INTERNAL_SERVER_ERROR));
   }
+
+
+
 };
 
 exports.postFileCategory = async (req, res, next) => {
